+++ datafusion/sqllogictest/Cargo.toml
@@ -49,7 +49,6 @@ indicatif = "0.17"
 itertools = { workspace = true }
 log = { workspace = true }
 object_store = { workspace = true }
-once_cell = { version = "1.20", optional = true }
 postgres-protocol = { version = "0.6.7", optional = true }
 postgres-types = { version = "0.2.8", features = ["derive", "with-chrono-0_4"], optional = true }
 rust_decimal = { version = "1.36.0", features = ["tokio-pg"] }
@@ -69,7 +68,6 @@ avro = ["datafusion/avro"]
 postgres = [
     "bytes",
     "chrono",
-    "once_cell",
     "postgres-types",
     "postgres-protocol",
     "testcontainers",
